requirements.txt
hydrogram @ https://github.com/hydrogram/hydrogram/archive/dev.zip
tgcrypto
asyncio_dgram
uvloop
loguru
pyyaml
